Refactor airports into Site directory/namespace
