+++ resources/views/site/airport/index.blade.php
@@ -10,14 +10,14 @@
                         <div class="col-md-6 text-center">
                             <ul class="list-unstyled">
                                 @foreach($airports->first() as $airport)
-                                    <li>@if($airport->major)<strong>@endif<a href="{{route('airport.view', $airport->icao)}}">{{$airport->name}} ({{$airport->icao}} / {{$airport->iata}})</a>@if($airport->major)</strong>@endif</li>
+                                    <li>@if($airport->major)<strong>@endif<a href="{{route('site.airport.view', $airport->icao)}}">{{$airport->name}} ({{$airport->icao}} / {{$airport->iata}})</a>@if($airport->major)</strong>@endif</li>
                                 @endforeach
                             </ul>
                         </div>
                         <div class="col-md-6 text-center">
                             <ul class="list-unstyled">
                                 @foreach($airports->last() as $airport)
-                                    <li>@if($airport->major)<strong>@endif<a href="{{route('airport.view', $airport->icao)}}">{{$airport->name}} ({{$airport->icao}} / {{$airport->iata}})</a>@if($airport->major)</strong>@endif</li>
+                                    <li>@if($airport->major)<strong>@endif<a href="{{route('site.airport.view', $airport->icao)}}">{{$airport->name}} ({{$airport->icao}} / {{$airport->iata}})</a>@if($airport->major)</strong>@endif</li>
                                 @endforeach
                             </ul>
                         </div>
